// src/app/com/annaniks/yev-yev/core/models/mesages.ts
export enum Messages {
    success = 'Գործողությունը հաջողությամբ ավարտված է',
    fail = 'Գործողությունը ձախողված է',
    failValidation = 'Պարտադիր դաշտերը լրացված չեն',
    photoNotAttached = 'Пожалуйста прикрепите изображение',
    photoAttachedSuccessfully = 'Изображение успешно прикреплен',
    wrongTimeRange = 'Неправильной интервал'
}